Moving FileOperations to Model
