+++ openstudiocore/src/model/Model_Common_Include.i
@@ -21,6 +21,8 @@
 //  #include <model/Relationship.hpp>
   #include <model/GenericModelObject.hpp>
   #include <model/GenericModelObject_Impl.hpp>
+  
+  #include <model/FileOperations.hpp>
 
   // central list of all concrete ModelObject header files (_Impl and non-_Impl)
   #include <model/ConcreteModelObjects.hpp>
