--- openstudiocore/src/openstudio_lib/FileOperations.cpp
@@ -1,757 +1,0 @@
-/**********************************************************************
-*  Copyright (c) 2008-2016, Alliance for Sustainable Energy.
-*  All rights reserved.
-*
-*  This library is free software; you can redistribute it and/or
-*  modify it under the terms of the GNU Lesser General Public
-*  License as published by the Free Software Foundation; either
-*  version 2.1 of the License, or (at your option) any later version.
-*
-*  This library is distributed in the hope that it will be useful,
-*  but WITHOUT ANY WARRANTY; without even the implied warranty of
-*  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
-*  Lesser General Public License for more details.
-*
-*  You should have received a copy of the GNU Lesser General Public
-*  License along with this library; if not, write to the Free Software
-*  Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
-**********************************************************************/
-
-#include "FileOperations.hpp"
-
-//#include "OSDocument.hpp"
-
-//#include "../runmanager/lib/RunManager.hpp"
-//#include "../runmanager/lib/RubyJobUtils.hpp"
-//#include "../runmanager/lib/WorkItem.hpp"
-//#include "../runmanager/lib/Workflow.hpp"
-//#include "../runmanager/lib/JobFactory.hpp"
-
-#include "../model/Model.hpp"
-#include "../model/WeatherFile.hpp"
-#include "../model/WeatherFile_Impl.hpp"
-
-#include "../osversion/VersionTranslator.hpp"
-
-//#include "../measure/OSArgument.hpp"
-
-#include "../energyplus/ReverseTranslator.hpp"
-
-#include "../utilities/filetypes/EpwFile.hpp"
-#include "../utilities/core/ApplicationPathHelpers.hpp"
-#include "../utilities/bcl/BCLMeasure.hpp"
-#include "../utilities/plot/ProgressBar.hpp"
-#include "../utilities/core/Logger.hpp"
-#include "../utilities/core/PathHelpers.hpp"
-#include "../utilities/core/Assert.hpp"
-
-#include <OpenStudio.hxx>
-
-#include <QDir>
-#include <QMessageBox>
-#include <QTemporaryFile>
-
-
-
-
-//#include "OSAppBase.hpp"
-
-namespace openstudio {
-
-  bool removeDir(const QString &dirName)
-  {
-    bool result = true;
-    
-    LOG_FREE(Info, "removeDir", "removeDir '" << toString(dirName) << "'");
-
-    QDir dir(dirName);
-    if (dir.exists(dirName)) {
-      bool test = true;
-
-      for (const QFileInfo& info : dir.entryInfoList(QDir::NoDotAndDotDot | QDir::System | QDir::Hidden  | QDir::AllDirs | QDir::Files, QDir::DirsFirst))
-      {
-        if (info.isDir()) {
-          test = removeDir(info.absoluteFilePath());
-          if (!test){
-            LOG_FREE(Error, "removeDir", "Failed to remove directory '" << toString(info.absoluteFilePath()) << "'");
-            result = false;
-
-            // DLM: do we really want to give up here?
-            //return result;
-          }
-        }else{
-          test = QFile::remove(info.absoluteFilePath());
-          if (!test){
-            LOG_FREE(Error, "removeDir", "Failed to remove file '" << toString(info.absoluteFilePath()) << "'");
-            result = false;
-
-            // DLM: do we really want to give up here?
-            //return result;
-          }
-        }
-      }
-      test = dir.rmdir(dirName);
-      if (!test){
-        LOG_FREE(Error, "removeDir", "Failed to remove directory '" << toString(dirName) << "'");
-        result = false;
-
-        // DLM: do we really want to give up here?
-        //return result;
-      }
-    }
-    
-    return result;
-  }
-
-  bool synchDirStructures(const QString &srcPath, const QString &dstPath)
-  {
-    bool result = true;
-    bool test = true;
-
-    QDir srcDir(srcPath);
-    QDir dstDir(dstPath);
-
-    if (srcDir.canonicalPath() == dstDir.canonicalPath()){
-      LOG_FREE(Warn, "synchDirStructures", "Cannot synch destination '" << toString(dstPath) << "' with source '" << toString(srcPath) << "' because they resolve to the same location");
-      return true; // already synched
-    }
-
-    LOG_FREE(Info, "synchDirStructures", "Synching destination '" << toString(dstPath) << "' with source '" << toString(srcPath) << "'");
-
-    // remove all files in dst as well as any directories in dst that are not in src
-    for (const QFileInfo &dstItemInfo : dstDir.entryInfoList(QDir::Dirs | QDir::Files | QDir::NoDotAndDotDot))
-    {
-      QString srcItemPath = srcPath + "/" + dstItemInfo.fileName();
-      QString dstItemPath = dstPath + "/" + dstItemInfo.fileName();
-
-      QFileInfo srcItemInfo = QFileInfo(srcItemPath);
-
-      if (dstItemInfo.isDir()){
-
-        if (srcItemInfo.exists() && srcItemInfo.isDir()){
-
-          // directory also exists in source
-          test = synchDirStructures(srcItemPath, dstItemPath);
-          if (!test){
-            LOG_FREE(Error, "synchDirStructures", "Failed to synch destination '" << toString(dstItemPath) << "' with source '" << toString(srcItemPath) << "'");
-            result = false;
-
-            // DLM: do we really want to give up here?
-            //return result;
-          }
-
-        }else{
-
-          // directory does not exist in source
-          test = removeDir(dstItemPath);
-          if (!test){
-            LOG_FREE(Error, "synchDirStructures", "Failed to remove directory '" << toString(dstItemPath) << "'");
-            result = false;
-
-            // DLM: do we really want to give up here?
-            //return result;
-          }
-        }
-
-      }else if (dstItemInfo.isFile()){
-        test = QFile::remove(dstItemInfo.absoluteFilePath());
-        if (!test){
-          LOG_FREE(Error, "synchDirStructures", "Failed to remove file '" << toString(dstItemInfo.absoluteFilePath()) << "'");
-          result = false;
-
-          // DLM: do we really want to give up here?
-          //return result;
-        }
-      }
-    }
-
-    return result;
-  }
-
-  bool copyDir(const QString &srcPath, const QString &dstPath)
-  {
-    bool result = true;
-    bool test = true;
-
-    LOG_FREE(Info, "copyDir", "copyDir '" << toString(srcPath) << "' to '" << toString(dstPath) << "'");
-
-    // ensure directory exists
-    QFileInfo dstInfo(dstPath);
-    QString fileName = dstInfo.fileName();
-    if (!dstInfo.exists() || !dstInfo.isDir()){
-      QDir parentDstDir(QFileInfo(dstPath).path());
-      LOG_FREE(Info, "copyDir", "Creating directory named = '" << toString(fileName) << "' in parentDstDir = '" << toString(parentDstDir.path()) << "'");
-      if (!parentDstDir.mkpath(fileName)){
-        LOG_FREE(Error, "copyDir", "Failed to create directory = '" << toString(fileName) << "' in parentDstDir = '" << toString(parentDstDir.path()) << "'");
-        return false;
-      }
-    }
-
-    QDir srcDir(srcPath);
-
-    // remove all files in dst as well as any directories in dst that are not in src
-    test = synchDirStructures(srcPath, dstPath);
-    if (!test){
-      LOG_FREE(Error, "copyDir", "Failed to synch destination '" << toString(dstPath) << "' with source '" << toString(srcPath) << "'");
-      result = false;
-    }
-    
-    // copy all files in src to dst
-    for (const QFileInfo &srcItemInfo : srcDir.entryInfoList(QDir::Dirs | QDir::Files | QDir::NoDotAndDotDot))
-    {
-      QString srcItemPath = srcPath + "/" + srcItemInfo.fileName();
-      QString dstItemPath = dstPath + "/" + srcItemInfo.fileName();
-
-      QFileInfo dstItemInfo = QFileInfo(dstItemPath);
-
-      if (srcItemInfo.isDir()) {
-        test = copyDir(srcItemPath, dstItemPath);
-        if (!test){
-          LOG_FREE(Error, "copyDir", "Failed to copy directory '" << toString(srcItemPath) << "' to '" << toString(dstItemPath) << "'");
-          result = false;
-
-          // DLM: do we really want to give up here?
-          //return false;
-        }
-      } else if (srcItemInfo.isFile()) {
-        test = QFile::copy(srcItemPath, dstItemPath);
-        if (!test){
-          LOG_FREE(Error, "copyDir", "Failed to copy file '" << toString(srcItemPath) << "' to '" << toString(dstItemPath) << "'");
-          result = false;
-
-          // DLM: do we really want to give up here?
-          //return false;
-        }
-      } 
-    }
-
-    return result;
-  }
-
-  openstudio::path createModelTempDir()
-  {
-    openstudio::path result;
-    auto tempFile = new QTemporaryFile();
-
-    if( tempFile->open() )
-    {
-      QString path = longPathName(tempFile->fileName());
-
-      // must close tempFile so you can create directory with same name on Windows
-      delete tempFile;
-
-      QDir modelTempDir(path);
-      LOG_FREE(Info, "createModelTempDir", "Creating directory '" << toString(modelTempDir.path()) << "'");
-      bool test = modelTempDir.mkpath(modelTempDir.path());
-      OS_ASSERT(test);
-
-      result = toPath(path);
-
-      return result;
-    }
-
-    delete tempFile;
-     
-    OS_ASSERT(false);
-
-    return openstudio::path();
-  } 
-
-  bool initializeModelTempDir(const openstudio::path& osmPath, const openstudio::path& modelTempDir)
-  {
-    bool result = true;
-    
-    if( osmPath.empty() || !openstudio::filesystem::exists(osmPath)){
-     
-      LOG_FREE(Debug, "initializeModelTempDir", "OSM path '" << toString(osmPath) << "' is empty or does not exist");
-      result = false;
-
-    }else{
-      openstudio::path fileName = osmPath.filename();
-      openstudio::path destination = modelTempDir / fileName;
-      LOG_FREE(Debug, "initializeModelTempDir", "Copying '" << toString(osmPath) << "' to '" << toString(destination) << "'");
-   
-      // copy osm file
-      bool test = QFile::copy(toQString(osmPath), toQString(destination));
-      if (!test){
-        LOG_FREE(Error, "initializeModelTempDir", "Could not copy '" << toString(osmPath) << "' to '" << toString(destination) << "'");
-      }
-
-      // Copy all files from existing resources dir into temp dir when opening
-      openstudio::path sourceDir = osmPath.parent_path() / osmPath.stem();
-      openstudio::path destDir = modelTempDir / toPath("resources");
-      if (openstudio::filesystem::exists(sourceDir)){
-        LOG_FREE(Debug, "initializeModelTempDir", "Copying '" << toString(sourceDir) << "' to '" << toString(destDir) << "'");
-
-        test = copyDir(toQString(sourceDir), toQString(destDir));
-        if (!test){
-          LOG_FREE(Error, "initializeModelTempDir", "Could not copy '" << toString(sourceDir) << "' to '" << toString(destDir) << "'");
-          result = false;
-        }
-      }
-    }
-
-    return result;
-  }
-
-  bool updateModelTempDir(openstudio::model::Model& model, const openstudio::path& modelTempDir)
-  {
-    bool modified = false;
-
-    // Create folder layout
-    openstudio::path runpath = modelTempDir / openstudio::toPath("resources/run");
-    openstudio::path scriptspath = modelTempDir / openstudio::toPath("resources/scripts");
-    openstudio::path filespath = modelTempDir / openstudio::toPath("resources/files");
-
-    if (!openstudio::filesystem::exists(runpath)) 
-    {
-      openstudio::filesystem::create_directories(runpath);
-      modified = true;
-    }
-
-    if (!openstudio::filesystem::exists(scriptspath)) 
-    {
-      openstudio::filesystem::create_directories(scriptspath);
-      modified = true;
-    }
-
-    if (!openstudio::filesystem::exists(filespath)) 
-    {
-      openstudio::filesystem::create_directories(filespath);
-      modified = true;
-    }
-
-    // weather file is fixed in OSDocument::fixWeatherFileOnOpen
-
-    return modified;
-  }
-
-  void saveModelTempDir(const openstudio::path& modelTempDir, const openstudio::path& osmPath)
-  {
-    bool test = true;
-
-    // must remove file, QFile::copy does not overwrite
-    QFileInfo osmInfo(toQString(osmPath));
-    if (osmInfo.exists() && osmInfo.isFile()){
-      test = QFile::remove(toQString(osmPath));
-      if (!test){
-        LOG_FREE(Error, "saveModelTempDir", "Could not remove previous osm at '" << toString(osmPath) << "'");
-      }
-    }
-
-    // copy osm file
-    openstudio::path srcPath = modelTempDir / osmPath.filename();
-    test = QFile::copy(toQString(srcPath), toQString(osmPath));
-    if (!test){
-      LOG_FREE(Error, "saveModelTempDir", "Could not copy osm from '" << toString(srcPath) << "' to '" << toString(osmPath) << "'");
-    }
-
-    if( QFileInfo(toQString(osmPath)).exists() ) {
-
-      // copy resources
-      openstudio::path srcDir = modelTempDir / toPath("resources");
-      openstudio::path dstDir =  osmPath.parent_path() / osmPath.stem();
-
-      LOG_FREE(Debug, "saveModelTempDir", "Copying " << toString(srcDir) << " to " << toString(dstDir));
-
-      test = copyDir(toQString(srcDir), toQString(dstDir));
-      if (!test){
-        LOG_FREE(Error, "saveModelTempDir", "Could not copy '" << toString(srcDir) << "' to '" << toString(dstDir) << "'");
-      }
-
-    }
-  }
-
-  openstudio::path saveModel(openstudio::model::Model model, const openstudio::path& osmPath, const openstudio::path& modelTempDir)
-  {
-
-    openstudio::path modelPath = osmPath;
-
-    if (getFileExtension(osmPath).empty()) {
-      modelPath = setFileExtension(osmPath,modelFileExtension(),false,true);
-    }
-
-    // save osm to temp directory, saveModelTempDir will copy to real location
-    openstudio::path tempModelPath = modelTempDir / modelPath.filename();
-    model.save(tempModelPath,true); 
-
-    LOG_FREE(Debug, "saveModel", "Saved model to '" << toString(tempModelPath) << "'");
-
-    saveModelTempDir(modelTempDir, modelPath);
-
-    return modelPath;
-  }
-
-  void removeModelTempDir(const openstudio::path& modelTempDir)
-  {
-    removeDir(toQString(modelTempDir));
-  }
-
-  bool saveRunManagerDatabase(const openstudio::path& osmPath, const openstudio::path& modelTempDir,
-                              std::vector <double> useRadianceForDaylightingCalculations,
-                              QWidget* parent)
-  {
-    return saveRunManagerDatabase(osmPath, modelTempDir, std::map<openstudio::path,std::vector<measure::OSMeasureInfo> >(), 
-        useRadianceForDaylightingCalculations, parent);
-  }
-
-  bool saveRunManagerDatabase(const openstudio::path& osmPath, const openstudio::path& modelTempDir, 
-                              const std::map<openstudio::path,std::vector<measure::OSMeasureInfo> >& userScriptsByFolder, 
-                              std::vector <double> useRadianceForDaylightingCalculations,
-                              QWidget* parent)
-  {
-    bool newToolsFound = false;
-
-    openstudio::path rmdbPath = modelTempDir / toPath("resources/run.db");
-    openstudio::path scriptsDir = modelTempDir / toPath("resources/scripts");
-    openstudio::path simulationDir = openstudio::toPath("run");
-
-    try {
-      //// this run db exists just to hold the jobs to be loaded from later, it is not unpaused
-      //openstudio::runmanager::RunManager rm(rmdbPath, true, true, false);
-
-      //// this should not be necessary as rm was constructed with truncate set to true?
-      //rm.clearJobs();
-
-      //openstudio::runmanager::ConfigOptions co = rm.getConfigOptions();
-
-      //newToolsFound = co.findTools(true, true, true, true);
-      //rm.setConfigOptions(co);
-
-      //openstudio::runmanager::Workflow wf;
-
-      //bool ruby_installed = rm.getConfigOptions().getTools().getAllByName("ruby").tools().size() > 0;
-      //
-      //openstudio::path radiancePath;
-
-      //std::vector<openstudio::runmanager::ToolInfo> rad = rm.getConfigOptions().getTools().getAllByName("rad").tools();
-
-      //if (!rad.empty())
-      //{
-      //  radiancePath = rad.back().localBinPath.parent_path();
-      //} 
-
-      //bool ruby_jobs_skipped = wf.addStandardWorkflow(scriptsDir, ruby_installed, getOpenStudioRubyIncludePath(), 
-      //                                                userScriptsByFolder, 
-      //                                                useRadianceForDaylightingCalculations.size() > 0, 
-      //                                                radiancePath,
-      //                                                modelTempDir / toPath("resources"), true);
-      //if (ruby_jobs_skipped)
-      //{
-      //  QMessageBox::information(parent, 
-      //      "No Ruby Interpreter",
-      //      "Measures were provided but no ruby interpreter could be found. Scripts will not be run.",
-      //      QMessageBox::Ok);
-      //}
-
-      //openstudio::runmanager::Tools tools = rm.getConfigOptions().getTools();
-      //wf.add(tools);
-      //wf.addParam(runmanager::JobParam("flatoutdir"));
-
-      //openstudio::runmanager::Job j = wf.create(simulationDir, toPath("../in.osm"));
-
-      //rm.enqueue(j, true);
-
-    } catch (const std::exception &e) {
-      QMessageBox::information(parent, 
-          "Error Initiating Simulation",
-          e.what(),
-          QMessageBox::Ok);
-    }
-
-    return newToolsFound;
-  }
-
-  //void pauseRunManager(openstudio::runmanager::RunManager& rm)
-  //{
-  //  rm.setPaused(true);
-
-  //  std::vector<openstudio::runmanager::Job> jobs = rm.getJobs();
-
-  //  for (auto itr = jobs.begin();
-  //       itr != jobs.end();
-  //       ++itr)
-  //  {
-  //    itr->setCanceled(true);
-  //  }
-
-  //  for (auto itr = jobs.begin();
-  //       itr != jobs.end();
-  //       ++itr)
-  //  { 
-  //    itr->requestStop();
-  //  }
-
-  //  for (auto itr = jobs.begin();
-  //       itr != jobs.end();
-  //       ++itr)
-  //  {
-  //    itr->waitForFinished();
-  //  }
-  //}
-
-  //void startRunManager(openstudio::runmanager::RunManager& rm, const openstudio::path& osmPath, const openstudio::path& modelTempDir,
-  //                     std::vector <double> useRadianceForDaylightingCalculations, bool requireCalibrationReports, QWidget* parent)
-  //{
-////    openstudio::path rmdbPath = modelTempDir / toPath("resources/run.db");
-  //  openstudio::path simulationDir = toPath("run");
-
-  //  try {
-
-  //    // disconnect signals from current jobs before clearing jobs
-  //    std::vector<openstudio::runmanager::Job> jobs = rm.getJobs();
-
-  //    for (auto itr = jobs.begin();
-  //         itr != jobs.end();
-  //         ++itr)
-  //    {
-  //      itr->disconnect();
-  //    }
-
-  //    rm.setPaused(true);
-
-  //    // clear current jobs
-  //    rm.clearJobs();
-
-  //    if (boost::filesystem::exists(modelTempDir / toPath("resources") / simulationDir)) {
-  //      boost::filesystem::remove_all(modelTempDir / toPath("resources") / simulationDir);
-  //    }
-  //    
-  //    boost::optional<analysisdriver::SimpleProject> p = OSAppBase::instance()->project();
-  //    if (p)
-  //    {
-  //      openstudio::runmanager::ConfigOptions co(true);
-
-  //      analysis::Problem prob = p->analysis().problem();
-  //      analysisdriver::AnalysisRunOptions runOptions = standardRunOptions(*p);
-  //      std::vector<runmanager::WorkItem> workitems(prob.createWorkflow(p->baselineDataPoint(), runOptions.rubyIncludeDirectory()).toWorkItems());
-
-  //      openstudio::BCLMeasure reportRequestMeasure = openstudio::BCLMeasure::reportRequestMeasure();
-  //      openstudio::BCLMeasure standardReportsMeasure = openstudio::BCLMeasure::standardReportMeasure();
-  //      openstudio::BCLMeasure calibrationReportsMeasure = openstudio::BCLMeasure::calibrationReportMeasure();
-
-  //      // DLM: always add this measure even if the user has their own copy, this is more clear
-  //      //bool standardReportsFound = findBCLMeasureWorkItem(workitems, standardReportsMeasure.uuid());
-  //      //if (!standardReportsFound){
-  //        bool test = addReportingMeasureWorkItem(workitems, standardReportsMeasure);
-  //        OS_ASSERT(test);
-  //      //}
-
-  //      // DLM: always add this measure even if the user has their own copy, this is more clear
-  //      //bool calibrationReportsFound = findBCLMeasureWorkItem(workitems, calibrationReportsMeasure.uuid());
-  //      //if (requireCalibrationReports && !calibrationReportsFound){
-  //      if (requireCalibrationReports){
-  //        bool test = addReportingMeasureWorkItem(workitems, calibrationReportsMeasure);
-  //        OS_ASSERT(test);
-  //      }
-  //      /*
-  //      // check if we need to use radiance
-  //      if (useRadianceForDaylightingCalculations)
-  //      {
-  //        std::vector<openstudio::runmanager::ToolInfo> rad = co.getTools().getAllByName("rad").tools();
-
-  //        if (!rad.empty())
-  //        {
-  //          openstudio::path radiancePath = rad.back().localBinPath.parent_path();
-
-  //          bool modeltoidffound(false);
-  //          for (auto itr = workitems.begin();
-  //              itr != workitems.end();
-  //              ++itr)
-  //          {
-  //            if (itr->type == openstudio::runmanager::JobType::ModelToIdf)
-  //            {
-  //              workitems.insert(itr, runmanager::Workflow::radianceDaylightCalculations(runOptions.rubyIncludeDirectory(), radiancePath));
-  //              modeltoidffound = true;
-  //              break;
-  //            }
-  //          }
-
-  //          OS_ASSERT(modeltoidffound);
-  //        } else {
-  //          QMessageBox::information(parent, 
-  //              "Error with initiating simulation.",
-  //              "Radiance simulation requested but radiance was not found.\nRadiance simulation will be skipped.",
-  //              QMessageBox::Ok);
-  //        }
-  //      }
-  //      */
-  //      // add the report request measure before energyplus but after any other energyplus measures
-  //      test = addReportRequestMeasureWorkItem(workitems, reportRequestMeasure);
-  //      OS_ASSERT(test);
-
-  //      runmanager::Workflow wf(workitems);
-  //      wf.add(co.getTools());
-  //      //openstudio::runmanager::JobParams params;
-  //      //params.append("cleanoutfiles", "none");
-  //      //wf.add(params);
-  //      runmanager::Job j = wf.create(modelTempDir / toPath("resources") / simulationDir, modelTempDir / openstudio::toPath("in.osm"));
-  //      runmanager::JobFactory::optimizeJobTree(j);
-  //      rm.enqueue(j, true);
-  //    }
-
-  //    // connect signals to new jobs
-  //    jobs = rm.getJobs();
-
-  //    for (auto itr = jobs.begin();
-  //        itr != jobs.end();
-  //        ++itr)
-  //    {
-  //      if (parent){
-  //        bool isConnected = itr->connect(SIGNAL(outputDataAdded(const openstudio::UUID &, const std::string &)), 
-  //                                        parent, SLOT(outputDataAdded(const openstudio::UUID &, const std::string &)));
-  //        OS_ASSERT(isConnected);
-  //      }
-
-  //      if (!itr->parent())
-  //      {
-  //        // need to reset the base path of the top level job so it can find its osm in the saved
-  //        // location
-  //        
-  //        // run in model dir
-  //        //itr->setBasePath(osmPath.parent_path() / osmPath.stem());
-
-  //        // run in temp dir
-  //        itr->setBasePath(modelTempDir / toPath("resources"));
-
-  //        if (parent){         
-  //          bool isConnected = itr->connect(SIGNAL(treeChanged(const openstudio::UUID &)), parent, SLOT(treeChanged(const openstudio::UUID &)));
-  //          OS_ASSERT(isConnected);
-  //        }
-
-  //        std::shared_ptr<OSDocument> currentDocument = OSAppBase::instance()->currentDocument();
-  //        if (currentDocument){         
-  //          bool isConnected = itr->connect(SIGNAL(treeChanged(const openstudio::UUID &)), 
-  //            currentDocument.get(), SIGNAL(treeChanged(const openstudio::UUID &)));
-  //          OS_ASSERT(isConnected);
-  //        }
-  //      }
-  //    }
-
-  //    rm.setPaused(false);
-
-  //  } catch (const std::exception &e) {
-  //    QMessageBox::information(parent, 
-  //        "Error Initiating Simulation",
-  //        e.what(),
-  //        QMessageBox::Ok);
-  //  }
-  //}
-
-  //bool findBCLMeasureWorkItem(const std::vector<runmanager::WorkItem>& workItems, const openstudio::UUID& uuid)
-  //{
-  //  for (auto itr = workItems.begin();
-  //       itr != workItems.end();
-  //       ++itr)
-  //  {
-  //    if (itr->type == openstudio::runmanager::JobType::UserScript)
-  //    {
-
-  //      try {
-  //        runmanager::JobParam parameters = itr->params.get("ruby_bclmeasureparameters");
-
-  //        for (std::vector<runmanager::JobParam>::const_iterator itr = parameters.children.begin();
-  //             itr != parameters.children.end();
-  //             ++itr)
-  //        {
-  //          if (itr->value == "bcl_measure_uuid"){
-  //            openstudio::UUID bclMeasureUUID = openstudio::toUUID(itr->children.at(0).value);
-  //            if (bclMeasureUUID == uuid){
-  //              return true;
-  //            }
-  //          }
-  //        }
-  //      } catch (const std::exception &) {
-  //      }
-  //    }
-  //  }
-
-  //  return false;
-  //}
-
-  //bool addReportRequestMeasureWorkItem(std::vector<runmanager::WorkItem>& workItems, const openstudio::BCLMeasure& bclMeasure)
-  //{
-  //  std::string reportingMeasureArgument = runmanager::getReportRequestMeasureArgument(workItems);
-
-  //  ruleset::OSArgument argument = ruleset::OSArgument::makeStringArgument("measures_json", true, false);
-  //  argument.setValue(reportingMeasureArgument);
-
-  //  std::vector<ruleset::OSArgument> arguments;
-  //  arguments.push_back(argument);
-
-  //  runmanager::RubyJobBuilder rubyJobBuilder(bclMeasure, arguments);
-
-  //  rubyJobBuilder.addInputFile(openstudio::runmanager::FileSelection("last"),
-  //                              openstudio::runmanager::FileSource("All"),
-  //                              ".*\\.idf",
-  //                              "in.idf");
-  //  rubyJobBuilder.addInputFile(openstudio::runmanager::FileSelection("last"),
-  //                              openstudio::runmanager::FileSource("All"),
-  //                              ".*\\.osm",
-  //                              "in.osm");
-
-  //  // be able to access OpenStudio Ruby bindings
-  //  rubyJobBuilder.setIncludeDir(getOpenStudioRubyIncludePath());
-
-
-  //  runmanager::WorkItem workItem = rubyJobBuilder.toWorkItem();
-
-  //  for (auto itr = workItems.begin();
-  //       itr != workItems.end();
-  //       ++itr)
-  //  {
-  //    // currently we are doing ExpandObjects->(E+ Measures)->EnergyPlusPreProcess->EnergyPlus
-  //    if (itr->type == openstudio::runmanager::JobType::EnergyPlusPreProcess)
-  //    {
-  //      workItems.insert(itr, workItem);
-  //      return true;
-  //    }
-  //  }
-
-  //  return false;
-  //}
-
-  //bool addReportingMeasureWorkItem(std::vector<runmanager::WorkItem>& workItems, const openstudio::BCLMeasure& bclMeasure)
-  //{
-  //  runmanager::RubyJobBuilder rubyJobBuilder(bclMeasure);
-
-  //  rubyJobBuilder.addInputFile(openstudio::runmanager::FileSelection("last"),
-  //                              openstudio::runmanager::FileSource("All"),
-  //                              ".*\\.idf",
-  //                              "in.idf");
-  //  rubyJobBuilder.addInputFile(openstudio::runmanager::FileSelection("last"),
-  //                              openstudio::runmanager::FileSource("All"),
-  //                              ".*\\.osm",
-  //                              "in.osm");
-
-  //  // be able to access OpenStudio Ruby bindings
-  //  rubyJobBuilder.setIncludeDir(getOpenStudioRubyIncludePath());
-
-  //  runmanager::WorkItem workItem = rubyJobBuilder.toWorkItem();
-
-  //  for (auto itr = workItems.begin();
-  //       itr != workItems.end();
-  //       ++itr)
-  //  {
-  //    if (itr->type == openstudio::runmanager::JobType::OpenStudioPostProcess)
-  //    {
-  //      workItems.insert(itr, workItem);
-  //      return true;
-  //    }
-  //  }
-
-  //  return false;
-  //}
-
-  boost::optional<openstudio::model::Model> modelFromOSM(const openstudio::path& path, openstudio::osversion::VersionTranslator& versionTranslator, openstudio::ProgressBar* progressBar)
-  {
-    return versionTranslator.loadModel(path, progressBar);
-  }
-
-  boost::optional<openstudio::model::Model> modelFromIDF(const openstudio::path& path, openstudio::energyplus::ReverseTranslator& reverseTranslator, openstudio::ProgressBar* progressBar)
-  { 
-    return reverseTranslator.loadModel(path, progressBar);
-  }
-
-} // openstudio
-
